MB-10533 Move a few rest/json related things to rest.py
This is a more useful place for gathering things we want to use for our rest requests
