+++ tasks/prime.py
@@ -20,9 +20,9 @@ from utils.constants import (
     SUPPORT_API_KEY,
     TEST_PDF,
     ZERO_UUID,
-    get_json_headers,
 )
 from utils.hosts import form_prime_path
+from utils.rest import get_json_headers
 from utils.types import JSONType
 from .base import CertTaskMixin, ParserTaskMixin, check_response
